# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This page is API documentation for all services relating common data or operation.",
        "title": "Web Order API",
        "contact": {
            "name": "Developer"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1/restaurant",
    "paths": {
        "/user/customer": {
            "post": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "ping example",
                "parameters": [
                    {
                        "description": "customer",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Base": {
            "type": "object",
            "properties": {
                "crt_at": {
                    "type": "string"
                },
                "upd_at": {
                    "type": "string"
                }
            }
        },
        "model.Customer": {
            "type": "object",
            "properties": {
                "base": {
                    "$ref": "#/definitions/model.Base"
                },
                "customer_address": {
                    "type": "string"
                },
                "customer_dob": {
                    "type": "string"
                },
                "customer_email": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "customer_phone": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                }
            }
        }
    }
}